// src/models/node-types/books/types/BookInput.ts
export type BookInput = {
    title: string | null | undefined 
    author: string | null | undefined 
    publisher: string | null | undefined 
    year_of_publication: number | null | undefined 
    isbn: string | null | undefined 
    pages: number | null | undefined 
    language: string | null | undefined 
}